+++ csaf-rs/src/csaf/aggregation/product_status/product_groups_by_id_map.rs
@@ -1,32 +1,13 @@
 use std::collections::hash_map::IntoIter;
-use std::collections::{HashMap, HashSet};
+use std::collections::HashMap;
 use std::ops::{Deref, DerefMut};
 
-use crate::csaf_traits::{ProductStatus, ProductStatusGroup, ProductStatusTrait};
+use crate::csaf_traits::{ProductStatusAndPath, ProductStatusGroup, ProductStatusTrait};
 
 /// Aggregation of product IDs, mapped to their [`ProductStatusGroup`], preserving
 /// the original [`ProductStatus`] and index each product originated from.
 #[derive(Debug, Clone)]
-pub struct ProductGroupsByIdMap(HashMap<String, HashSet<ProductStatusGroupAndPath>>);
-
-/// A [`ProductStatusGroup`] with its original [`ProductStatus`] and the index
-/// within the original status list.
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub struct ProductStatusGroupAndPath {
-    pub status_group: ProductStatusGroup,
-    pub status: ProductStatus,
-    pub index: usize,
-}
-
-impl ProductStatusGroupAndPath {
-    /// Returns the JSON path for this product status entry relative to a vulnerability.
-    pub fn json_path(&self, vulnerability_index: usize) -> String {
-        format!(
-            "/vulnerabilities/{}/product_status/{}/{}",
-            vulnerability_index, self.status, self.index
-        )
-    }
-}
+pub struct ProductGroupsByIdMap(HashMap<String, HashMap<ProductStatusGroup, Vec<ProductStatusAndPath>>>);
 
 impl<T: ProductStatusTrait> From<&T> for ProductGroupsByIdMap {
     /// Construct a [`ProductGroupsByIdMap`] from anything implementing [`ProductStatusTrait`].
@@ -40,8 +21,9 @@ impl<T: ProductStatusTrait> From<&T> for ProductGroupsByIdMap {
                     .0
                     .entry(product_id)
                     .or_default()
-                    .insert(ProductStatusGroupAndPath {
-                        status_group: group.clone(),
+                    .entry(group.clone())
+                    .or_default()
+                    .push(ProductStatusAndPath {
                         status: status.clone(),
                         index,
                     });
@@ -53,7 +35,7 @@ impl<T: ProductStatusTrait> From<&T> for ProductGroupsByIdMap {
 }
 
 impl Deref for ProductGroupsByIdMap {
-    type Target = HashMap<String, HashSet<ProductStatusGroupAndPath>>;
+    type Target = HashMap<String, HashMap<ProductStatusGroup, Vec<ProductStatusAndPath>>>;
 
     fn deref(&self) -> &Self::Target {
         &self.0
@@ -67,8 +49,8 @@ impl DerefMut for ProductGroupsByIdMap {
 }
 
 impl IntoIterator for ProductGroupsByIdMap {
-    type Item = (String, HashSet<ProductStatusGroupAndPath>);
-    type IntoIter = IntoIter<String, HashSet<ProductStatusGroupAndPath>>;
+    type Item = (String, HashMap<ProductStatusGroup, Vec<ProductStatusAndPath>>);
+    type IntoIter = IntoIter<String, HashMap<ProductStatusGroup, Vec<ProductStatusAndPath>>>;
 
     fn into_iter(self) -> Self::IntoIter {
         self.0.into_iter()
