+++ csaf-rs/src/csaf/aggregation/product_status/product_status_group_map.rs
@@ -1,33 +1,14 @@
 use std::collections::hash_map::IntoIter;
-use std::collections::{HashMap, HashSet};
+use std::collections::HashMap;
 use std::ops::{Deref, DerefMut};
 
-use crate::csaf::enums::product_status::ProductStatus;
+use crate::csaf::aggregation::product_status::ProductStatusAndPath;
 use crate::csaf_traits::{ProductStatusGroup, ProductStatusTrait};
 
 /// Aggregation of product IDs grouped by their [`ProductStatusGroup`], preserving
 /// the original [`ProductStatus`] and index each product originated from.
 #[derive(Debug, Clone)]
-pub struct ProductStatusGroupMap(HashMap<ProductStatusGroup, HashSet<ProductStatusAndPath>>);
-
-/// A product ID with its original [`ProductStatus`] and the index
-/// within the original status list.
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub struct ProductStatusAndPath {
-    pub product_id: String,
-    pub status: ProductStatus,
-    pub index: usize,
-}
-
-impl ProductStatusAndPath {
-    /// Returns the JSON path for this product status entry relative to a vulnerability.
-    pub fn json_path(&self, vulnerability_index: usize) -> String {
-        format!(
-            "/vulnerabilities/{}/product_status/{}/{}",
-            vulnerability_index, self.status, self.index
-        )
-    }
-}
+pub struct ProductStatusGroupMap(HashMap<ProductStatusGroup, HashMap<String, Vec<ProductStatusAndPath>>>);
 
 impl<T: ProductStatusTrait> From<&T> for ProductStatusGroupMap {
     /// Construct a [`ProductStatusGroupMap`] from anything implementing [`ProductStatusTrait`].
@@ -38,9 +19,8 @@ impl<T: ProductStatusTrait> From<&T> for ProductStatusGroupMap {
             let group = ProductStatusGroup::from(&status);
             let entry = result.0.entry(group).or_default();
             for (index, product_id) in products.into_iter().enumerate() {
-                entry.insert(ProductStatusAndPath {
+                entry.entry(product_id).or_default().push(ProductStatusAndPath {
                     status: status.clone(),
-                    product_id,
                     index,
                 });
             }
@@ -55,12 +35,12 @@ impl ProductStatusGroupMap {
     pub fn contains(&self, group: &ProductStatusGroup, product_id: &str) -> bool {
         self.0
             .get(group)
-            .is_some_and(|entries| entries.iter().any(|e| e.product_id == product_id))
+            .is_some_and(|entries| entries.contains_key(product_id))
     }
 }
 
 impl Deref for ProductStatusGroupMap {
-    type Target = HashMap<ProductStatusGroup, HashSet<ProductStatusAndPath>>;
+    type Target = HashMap<ProductStatusGroup, HashMap<String, Vec<ProductStatusAndPath>>>;
 
     fn deref(&self) -> &Self::Target {
         &self.0
@@ -74,8 +54,8 @@ impl DerefMut for ProductStatusGroupMap {
 }
 
 impl IntoIterator for ProductStatusGroupMap {
-    type Item = (ProductStatusGroup, HashSet<ProductStatusAndPath>);
-    type IntoIter = IntoIter<ProductStatusGroup, HashSet<ProductStatusAndPath>>;
+    type Item = (ProductStatusGroup, HashMap<String, Vec<ProductStatusAndPath>>);
+    type IntoIter = IntoIter<ProductStatusGroup, HashMap<String, Vec<ProductStatusAndPath>>>;
 
     fn into_iter(self) -> Self::IntoIter {
         self.0.into_iter()
