+++ csaf-rs/src/csaf_traits.rs
@@ -1,9 +1,6 @@
-pub use crate::csaf::aggregation::product_status::product_groups_by_id_map::{
-    ProductGroupsByIdMap, ProductStatusGroupAndPath,
-};
-pub use crate::csaf::aggregation::product_status::product_status_group_map::{
-    ProductStatusAndPath, ProductStatusGroupMap,
-};
+pub use crate::csaf::aggregation::product_status::product_groups_by_id_map::ProductGroupsByIdMap;
+pub use crate::csaf::aggregation::product_status::ProductStatusAndPath;
+pub use crate::csaf::aggregation::product_status::product_status_group_map::ProductStatusGroupMap;
 pub use crate::csaf::enums::category_of_the_branch::CategoryOfTheBranch;
 pub use crate::csaf::enums::csaf_version::CsafVersion;
 pub use crate::csaf::enums::product_status::ProductStatus;
